Move WrapperValue tests to new test files
